Make s3ObjectReader code more generic
