+++ go/store/nbs/aws_chunk_source.go
@@ -32,7 +32,7 @@ import (
 
 func tableExistsInChunkSource(ctx context.Context, s3 *s3ObjectReader, al awsLimits, name hash.Hash, chunkCount uint32, q MemoryQuotaProvider, stats *Stats) (bool, error) {
 	magic := make([]byte, magicNumberSize)
-	n, _, err := s3.ReadFromEnd(ctx, name, magic, stats)
+	n, err := readS3TableFileFromEnd(ctx, s3, name.String(), magic, stats)
 	if err != nil {
 		return false, err
 	}
@@ -42,10 +42,11 @@ func tableExistsInChunkSource(ctx context.Context, s3 *s3ObjectReader, al awsLim
 	return bytes.Equal(magic, []byte(magicNumber)), nil
 }
 
+// NM4 - Rename to newAWSTableFileChunkSource, maybe?
 func newAWSChunkSource(ctx context.Context, s3 *s3ObjectReader, al awsLimits, name hash.Hash, chunkCount uint32, q MemoryQuotaProvider, stats *Stats) (cs chunkSource, err error) {
 	var tra tableReaderAt
 	index, err := loadTableIndex(ctx, stats, chunkCount, q, func(p []byte) error {
-		n, _, err := s3.ReadFromEnd(ctx, name, p, stats)
+		n, err := readS3TableFileFromEnd(ctx, s3, name.String(), p, stats)
 		if err != nil {
 			return err
 		}
